Move builtin to status module namespace
