+++ cmd/fynedesk/main.go
@@ -4,7 +4,7 @@ import (
 	"fyne.io/fyne/app"
 	"fyne.io/fyne/theme"
 
-	_ "fyne.io/fynedesk/modules/builtin"
+	_ "fyne.io/fynedesk/modules/status"
 )
 
 func main() {
